Fixed AVCC and Big Button pinout on schamatic
The AVCC should be connected to VCC and not 3.3V because when it's powered by 5V the AVCC pin would output 5V towards the Voltage regulator. The SW2 (big button) had ground on the wrong pin (floating input). It's now corrected so that it properly acts as an active high button.